+++ src/jit/compiler.cpp
@@ -1423,7 +1423,7 @@ void                Compiler::compInit(norls_allocator * pAlloc, InlineInfo * in
     compMaxUncheckedOffsetForNullObject = MAX_UNCHECKED_OFFSET_FOR_NULL_OBJECT;
 
     compNativeSizeEstimate = NATIVE_SIZE_INVALID;
-    compInlineeHints = (InlInlineHints)0;
+    compInlineeHints = (InlineHints)0;
 
     for (unsigned i = 0; i < MAX_LOOP_NUM; i++)
     {
@@ -4970,7 +4970,7 @@ int           Compiler::compCompileHelper (CORINFO_MODULE_HANDLE            clas
             if (opts.eeFlags & CORJIT_FLG_PREJIT)
             {
                 // Cache inlining hint during NGen to avoid touching bodies of non-inlineable methods at runtime
-                JitInlineResult trialResult(this, nullptr, methodHnd, "prejit1");
+                InlineResult trialResult(this, nullptr, methodHnd, "prejit1");
                 impCanInlineIL(methodHnd, methodInfo, forceInline, &trialResult);
                 if (trialResult.isFailure())
                 {
@@ -5016,7 +5016,7 @@ int           Compiler::compCompileHelper (CORINFO_MODULE_HANDLE            clas
             assert(compNativeSizeEstimate != NATIVE_SIZE_INVALID); 
 
             int callsiteNativeSizeEstimate = impEstimateCallsiteNativeSize(methodInfo);
-            JitInlineResult trialResult(this, nullptr, methodHnd, "prejit2");
+            InlineResult trialResult(this, nullptr, methodHnd, "prejit2");
             
             impCanInlineNative(callsiteNativeSizeEstimate, 
                                compNativeSizeEstimate,
@@ -9832,49 +9832,3 @@ HelperCallProperties Compiler::s_helperCallProperties;
 
 /*****************************************************************************/
 /*****************************************************************************/
-
-//------------------------------------------------------------------------
-// report: Dump, log, and report information about an inline decision.
-//
-// Notes:
-//    
-//    Called (automatically via the JitInlineResult dtor) when the inliner
-//    is done evaluating a candidate.
-//
-//    Dumps state of the inline candidate, and if a decision was reached
-//    sends it to the log and reports the decision back to the EE. 
-//    
-//    All this can be suppressed if desired by calling setReported() before 
-//    the JitInlineResult goes out of scope.
-
-void JitInlineResult::report() 
-{
-    // User may have suppressed reporting via setReported(). If so, do nothing.
-    if (inlReported)
-    {
-        return;
-    }
-
-    inlReported = true;
-
-#ifdef DEBUG
-
-    if (VERBOSE)
-    {
-        const char* format = "INLINER: during '%s' result '%s' reason '%s' for '%s' calling '%s'\n";
-        const char* caller = (inlInliner == nullptr) ? "n/a" : inlCompiler->eeGetMethodFullName(inlInliner);
-        const char* callee = (inlInlinee == nullptr) ? "n/a" : inlCompiler->eeGetMethodFullName(inlInlinee);
-
-        JITDUMP(format, inlContext, resultString(), inlReason, caller, callee);
-    }
-
-#endif // DEBUG
-
-    if (isDecided()) 
-    {
-        const char* format = "INLINER: during '%s' result '%s' reason '%s'\n";
-        JITLOG_THIS(inlCompiler, (LL_INFO100000, format, inlContext, resultString(), inlReason));
-        COMP_HANDLE comp = inlCompiler->info.compCompHnd;
-        comp->reportInliningDecision(inlInliner, inlInlinee, result(), inlReason);
-    }
-}
